+++ CHANGES.md
@@ -1,4 +1,5 @@
 [unreleased]
+* add filter `gu_running_git_servers`
 
 #### 0.7.0 / 2021-03-15 🎂
 * add filter `gu_get_git_icon_data`